achx: remove 'remoteach' sub-package
