+++ internal/achx/ppd.go
@@ -2,58 +2,55 @@
 // Use of this source code is governed by an Apache License
 // license that can be found in the LICENSE file.
 
-package remoteach
+package achx
 
 import (
 	"errors"
 	"fmt"
 
 	"github.com/moov-io/ach"
-	"github.com/moov-io/paygate/internal/achx"
 	"github.com/moov-io/paygate/internal/model"
 )
 
-func createCCDBatch(id string, transfer *model.Transfer, receiver *model.Receiver, receiverDep *model.Depository, orig *model.Originator, origDep *model.Depository) (ach.Batcher, error) {
-	if transfer.CCDDetail == nil {
-		return nil, errors.New("nil CCD detail")
-	}
-	if transfer.CCDDetail.PaymentInformation == "" {
-		return nil, fmt.Errorf("transfer=%s CCD transfer is missing PaymentInformation", id)
-	}
-
+func createPPDBatch(id string, transfer *model.Transfer, receiver *model.Receiver, receiverDep *model.Depository, orig *model.Originator, origDep *model.Depository) (ach.Batcher, error) {
 	batchHeader := makeBatchHeader(id, transfer, orig, origDep)
-	batchHeader.StandardEntryClassCode = ach.CCD
+	batchHeader.StandardEntryClassCode = ach.PPD
 
-	// Add EntryDetail to CCD batch
+	// Add EntryDetail to PPD batch
 	entryDetail := ach.NewEntryDetail()
 	entryDetail.ID = id
 	entryDetail.TransactionCode = determineTransactionCode(transfer, origDep)
-	entryDetail.RDFIIdentification = achx.ABA8(receiverDep.RoutingNumber)
-	entryDetail.CheckDigit = achx.ABACheckDigit(receiverDep.RoutingNumber)
+	entryDetail.RDFIIdentification = ABA8(receiverDep.RoutingNumber)
+	entryDetail.CheckDigit = ABACheckDigit(receiverDep.RoutingNumber)
 	entryDetail.Amount = transfer.Amount.Int()
 	entryDetail.IdentificationNumber = createIdentificationNumber()
 	entryDetail.IndividualName = receiver.Metadata
-	entryDetail.TraceNumber = achx.TraceNumber(origDep.RoutingNumber)
+	entryDetail.DiscretionaryData = transfer.Description
+	entryDetail.TraceNumber = TraceNumber(origDep.RoutingNumber)
 
 	if num, err := receiverDep.DecryptAccountNumber(); err != nil {
-		return nil, fmt.Errorf("CCD: receiver account number decrypt failed: %v", err)
+		return nil, fmt.Errorf("PPD: receiver account number decrypt failed: %v", err)
 	} else {
 		entryDetail.DFIAccountNumber = num
+	}
+
+	if transfer.PPDDetail == nil {
+		return nil, errors.New("nil PPD detail")
 	}
 
 	// Add Addenda05
 	addenda05 := ach.NewAddenda05()
 	addenda05.ID = id
-	addenda05.PaymentRelatedInformation = transfer.CCDDetail.PaymentInformation
+	addenda05.PaymentRelatedInformation = transfer.PPDDetail.PaymentInformation
 	addenda05.SequenceNumber = 1
 	addenda05.EntryDetailSequenceNumber = 1
 	entryDetail.AddAddenda05(addenda05)
 	entryDetail.AddendaRecordIndicator = 1
 
-	// For now just create CCD
+	// For now just create PPD
 	batch, err := ach.NewBatch(batchHeader)
 	if err != nil {
-		return nil, fmt.Errorf("failed to create CCD batch: %v", err)
+		return nil, fmt.Errorf("failed to create PPD batch: %v", err)
 	}
 	batch.AddEntry(entryDetail)
 	batch.SetControl(ach.NewBatchControl())
