+++ internal/achx/tel_test.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by an Apache License
 // license that can be found in the LICENSE file.
 
-package remoteach
+package achx
 
 import (
 	"encoding/json"
@@ -15,8 +15,8 @@ import (
 	"github.com/moov-io/paygate/pkg/id"
 )
 
-func TestWEBPaymentType(t *testing.T) {
-	var paymentType model.WEBPaymentType
+func TestTELPaymentType(t *testing.T) {
+	var paymentType model.TELPaymentType
 	if err := json.Unmarshal([]byte(`"SINGLE"`), &paymentType); err != nil {
 		t.Fatal(err)
 	}
@@ -24,14 +24,14 @@ func TestWEBPaymentType(t *testing.T) {
 		t.Fatal(err)
 	}
 	if err := json.Unmarshal([]byte(`"other"`), &paymentType); err == nil {
-		t.Fatal(err)
+		t.Fatal("expected error")
 	}
 	if err := json.Unmarshal([]byte("1"), &paymentType); err == nil {
 		t.Fatal("expected error")
 	}
 }
 
-func TestWEB__createWEBBatch(t *testing.T) {
+func TestTEL__createTELBatch(t *testing.T) {
 	depID, userID := base.ID(), id.User(base.ID())
 	keeper := secrets.TestStringKeeper(t)
 
@@ -62,15 +62,16 @@ func TestWEB__createWEBBatch(t *testing.T) {
 		Metadata:          "jane doe",
 	}
 	origDep := &model.Depository{
-		ID:            id.Depository(base.ID()),
-		BankName:      "foo bank",
-		Holder:        "john doe",
-		HolderType:    model.Individual,
-		Type:          model.Savings,
-		RoutingNumber: "231380104",
-		Status:        model.DepositoryVerified,
-		Metadata:      "john doe savings",
-		Keeper:        keeper,
+		ID:                     id.Depository(base.ID()),
+		BankName:               "foo bank",
+		Holder:                 "john doe",
+		HolderType:             model.Individual,
+		Type:                   model.Savings,
+		RoutingNumber:          "231380104",
+		EncryptedAccountNumber: "2",
+		Status:                 model.DepositoryVerified,
+		Metadata:               "john doe savings",
+		Keeper:                 keeper,
 	}
 	origDep.ReplaceAccountNumber("2")
 	orig := &model.Originator{
@@ -89,21 +90,20 @@ func TestWEB__createWEBBatch(t *testing.T) {
 		Receiver:               receiver.ID,
 		ReceiverDepository:     receiverDep.ID,
 		Description:            "sending money",
-		StandardEntryClassCode: "WEB",
+		StandardEntryClassCode: "TEL",
 		Status:                 model.TransferPending,
 		UserID:                 userID.String(),
-		WEBDetail: &model.WEBDetail{
-			PaymentInformation: "test payment",
-			PaymentType:        model.WEBSingle,
+		TELDetail: &model.TELDetail{
+			PaymentType: "single",
 		},
 	}
 
-	batch, err := createWEBBatch(depID, transfer, receiver, receiverDep, orig, origDep)
+	batch, err := createTELBatch(depID, transfer, receiver, receiverDep, orig, origDep)
 	if err != nil {
 		t.Fatal(err)
 	}
 	if batch == nil {
-		t.Error("nil WEB Batch")
+		t.Error("nil TEL Batch")
 	}
 
 	file, err := ConstructFile(depID, "", gateway, transfer, receiver, receiverDep, orig, origDep)
@@ -111,16 +111,16 @@ func TestWEB__createWEBBatch(t *testing.T) {
 		t.Fatal(err)
 	}
 	if file == nil {
-		t.Error("nil WEB ach.File")
+		t.Error("nil TEL ach.File")
 	}
 
-	// Make sure WEBReoccurring are rejected
-	transfer.WEBDetail.PaymentType = "reoccurring"
-	batch, err = createWEBBatch(depID, transfer, receiver, receiverDep, orig, origDep)
+	// Make sure TELReoccurring are rejected
+	transfer.TELDetail.PaymentType = "reoccurring"
+	batch, err = createTELBatch(depID, transfer, receiver, receiverDep, orig, origDep)
 	if batch != nil || err == nil {
 		t.Errorf("expected error, but got batch: %v", batch)
 	} else {
-		if !strings.Contains(err.Error(), "createWEBBatch: reoccurring WEB transfers are not supported") {
+		if !strings.Contains(err.Error(), "createTELBatch: reoccurring TEL transfers are not supported") {
 			t.Fatalf("unexpected error: %v", err)
 		}
 	}
