+++ internal/achx/web_test.go
@@ -2,9 +2,11 @@
 // Use of this source code is governed by an Apache License
 // license that can be found in the LICENSE file.
 
-package remoteach
+package achx
 
 import (
+	"encoding/json"
+	"strings"
 	"testing"
 
 	"github.com/moov-io/base"
@@ -13,7 +15,23 @@ import (
 	"github.com/moov-io/paygate/pkg/id"
 )
 
-func TestPPD__createPPDBatch(t *testing.T) {
+func TestWEBPaymentType(t *testing.T) {
+	var paymentType model.WEBPaymentType
+	if err := json.Unmarshal([]byte(`"SINGLE"`), &paymentType); err != nil {
+		t.Fatal(err)
+	}
+	if err := json.Unmarshal([]byte(`"ReoCCuRRing"`), &paymentType); err != nil {
+		t.Fatal(err)
+	}
+	if err := json.Unmarshal([]byte(`"other"`), &paymentType); err == nil {
+		t.Fatal(err)
+	}
+	if err := json.Unmarshal([]byte("1"), &paymentType); err == nil {
+		t.Fatal("expected error")
+	}
+}
+
+func TestWEB__createWEBBatch(t *testing.T) {
 	depID, userID := base.ID(), id.User(base.ID())
 	keeper := secrets.TestStringKeeper(t)
 
@@ -71,20 +89,21 @@ func TestPPD__createPPDBatch(t *testing.T) {
 		Receiver:               receiver.ID,
 		ReceiverDepository:     receiverDep.ID,
 		Description:            "sending money",
-		StandardEntryClassCode: "PPD",
+		StandardEntryClassCode: "WEB",
 		Status:                 model.TransferPending,
 		UserID:                 userID.String(),
-		PPDDetail: &model.PPDDetail{
-			PaymentInformation: "payment",
+		WEBDetail: &model.WEBDetail{
+			PaymentInformation: "test payment",
+			PaymentType:        model.WEBSingle,
 		},
 	}
 
-	batch, err := createPPDBatch(depID, transfer, receiver, receiverDep, orig, origDep)
+	batch, err := createWEBBatch(depID, transfer, receiver, receiverDep, orig, origDep)
 	if err != nil {
 		t.Fatal(err)
 	}
 	if batch == nil {
-		t.Error("nil PPD Batch")
+		t.Error("nil WEB Batch")
 	}
 
 	file, err := ConstructFile(depID, "", gateway, transfer, receiver, receiverDep, orig, origDep)
@@ -92,6 +111,17 @@ func TestPPD__createPPDBatch(t *testing.T) {
 		t.Fatal(err)
 	}
 	if file == nil {
-		t.Error("nil PPD ach.File")
+		t.Error("nil WEB ach.File")
+	}
+
+	// Make sure WEBReoccurring are rejected
+	transfer.WEBDetail.PaymentType = "reoccurring"
+	batch, err = createWEBBatch(depID, transfer, receiver, receiverDep, orig, origDep)
+	if batch != nil || err == nil {
+		t.Errorf("expected error, but got batch: %v", batch)
+	} else {
+		if !strings.Contains(err.Error(), "createWEBBatch: reoccurring WEB transfers are not supported") {
+			t.Fatalf("unexpected error: %v", err)
+		}
 	}
 }
